refactor(home): extract weekly statistics service
